Move load of mods to inside the list
Since this will be reused across the site I might as well make the load a part of it. Gotta wait for nuxt3 to have it work better as the current fetch sucks and doesn't work well in SSR and AsyncData doesn't work in non pages. Nuxt3 afaik will have AsyncData that works across all components.